uncomented setup controller lines

+++ api/controller/masterController.js
@@ -17,11 +17,11 @@ const Logins = require(appRoot + "/models/loginRegistryModel");
 module.exports = function(app){
 
     app.get("/setup", function(req, res){
-         /* if(env !== "development" && env !== "pi"){
+          if(env !== "development" && env !== "pi"){
             logger.warn(req.connection.remoteAddress + " tried to access /setup in " + env);
             res.send("setup not available at this environment");
           }
-          else{*/
+          else{
             logger.warn("database is about to be dropped");
             mongoose.connection.db.dropDatabase();
             createUsers()
@@ -33,7 +33,7 @@ module.exports = function(app){
                     logger.error("err: " + err);
                     return res.send({success:false, msg:err});
                 });
-      /*    }*/
+         }
     });
 
      /**
